jdbc updated

+++ JavaCoreDev/src/main/java/interfaces/Switchable.java
@@ -3,5 +3,5 @@ package main.java.interfaces;
 import main.java.exceptions.NonCorrectModeException;
 
 public interface Switchable {
-    void setMode(byte mode) throws NonCorrectModeException;
+    void setMode(int mode) throws NonCorrectModeException;
 }
